#642 rename wbs details page folder
