+++ src/client/pages/WbsDetailsSwitchPage/wbs-details.module.css
