+++ src/client/pages/WbsDetailsSwitchPage/wbs-details.test.tsx
