+++ src/client/pages/WbsDetailsSwitchPage/wbs-details.tsx
